Update type import and export
